backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
openai>=1.0
python-dotenv>=1.0
psycopg2-binary>=2.9
